fix errors + add tests

+++ src/services/tasks.service.ts
@@ -34,10 +34,16 @@ export const addTask = async (task: string): Promise<string> => {
 
 export const deleteTaskByID = async (id: string): Promise<void> => {
     try{
-        await getTaskByID(id);
+        const task = await getTaskByID(id);
+    
+        if(task.id === undefined){
+            throw(task);
+        }
+        
         await Task.destroy({
             where:{ id }
         });
+
     }catch(err){
         return(err);
     }
@@ -52,6 +58,11 @@ export const deleteTaskByStatus = async (status: string): Promise<void> => {
 export const updateTask = async (id: string): Promise<void> => {
     try{
         const task = await getTaskByID(id);
+
+        if(task.status === undefined){
+            throw(task);
+        }
+
         if(task.status === ETaskStatus.Incomplete){
             await Task.update({status: ETaskStatus.Complete},{
                 where:{ id }
@@ -60,8 +71,7 @@ export const updateTask = async (id: string): Promise<void> => {
             await Task.update({status: ETaskStatus.Incomplete},{
                 where:{ id }
             });
-        }
-        
+        } 
     }catch(err){
         return(err);
     }
